requirements.txt
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
az-iranian-bank-gateways==1.7.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
black==22.10.0
bleach==5.0.1
blinker==1.4
branca==0.5.0
Brlapi==0.8.3
Brotli==1.0.9
cached-property==1.5.2
celery==5.2.7
certifi==2022.6.15.1
cffi==1.15.1
chardet==4.0.0
chromedriver-py==107.0.5304.62
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
command-not-found==0.3
commonmark==0.9.1
ConfigArgParse==1.5.3
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.2
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==4.0
django-adminlte-3==0.1.6
django-celery-beat==2.4.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-cuser==2017.3.16
django-filter==22.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-phonenumber-field==7.0.0
django-taggit==3.1.0
django-timezone-field==5.0
django-tinymce==3.5.0
django-treebeard==4.5.1
djangorestframework==3.14.0
draftjs-exporter==2.1.7
duplicity==0.8.21
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.10.0
fasteners==0.14.1
fastjsonschema==2.16.1
filelock==3.7.1
flake8==5.0.3
Flask==2.1.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
folium==0.12.1.post1
future==0.18.2
gevent==21.12.0
geventhttpclient==2.0
gpg===1.16.0-unknown
greenlet==1.1.2
h11==0.13.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
isodate==0.6.1
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.14.0
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
kavenegar==1.1.2
keyring==23.5.0
Khayyam==3.0.17
kombu==5.2.4
l18n==2021.3
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
locust==2.10.2
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mime==0.1.0
mistune==0.8.4
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.4
mypy-extensions==0.4.3
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.5
netifaces==0.11.0
nltk==3.7
notebook==6.4.12
numpy==1.23.2
oauthlib==3.2.0
olefile==0.46
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
phonenumbers==8.12.57
phonenumberslite==8.12.57
pickleshare==0.7.5
pika==1.3.0
Pillow==9.0.1
platformdirs==2.5.2
pluggy==1.0.0
pool==0.1.2.dev0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.12.4
psutil==5.9.1
psycopg-binary==3.1.8
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycairo==1.20.1
pycodestyle==2.9.0
pycparser==2.21
pycryptodome==3.15.0
pycups==2.0.1
pyflakes==2.5.0
Pygments==2.12.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymemcache==4.0.0
pymongo==4.2.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
python-apt==2.3.0+ubuntu2.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-gettext==4.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==23.2.1
qtconsole==5.3.1
QtPy==2.2.0
redis==4.3.4
regex==2022.10.31
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.10.0
restapi3==0.0.1
rich==12.5.1
roundrobin==0.0.2
scipy==1.9.3
SecretStorage==3.3.1
selenium==4.4.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.4.0
suds==1.1.2
systemd-python==234
telepath==0.3
terminado==0.15.0
tinycss2==1.1.1
tomli==2.0.1
tornado==6.2
tox==3.27.0
tqdm==4.64.1
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
tzdata==2022.6
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.15.1
wadllib==1.3.6
wagtail==4.1.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.1
widgetsnbextension==4.0.2
Willow==1.4.1
wrapt==1.14.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
zeep==4.1.0
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
